backend/requirements.txt
Flask
Flask-CORS
Flask-MySQLdb
werkzeug
python-dotenv